chore: move some helper to core
