--- packages/shared/src/createHook.ts
@@ -1,53 +1,0 @@
-import { isFunction } from './utils';
-
-type HookOrder = 'pre' | 'post' | 'default';
-
-export type HookDescriptor<T extends (...args: any[]) => any> = {
-  handler: T;
-  order: HookOrder;
-};
-
-export type AsyncHook<Callback extends (...args: any[]) => any> = {
-  tap: (cb: Callback | HookDescriptor<Callback>) => void;
-  call: (...args: Parameters<Callback>) => Promise<Parameters<Callback>>;
-};
-
-export function createAsyncHook<
-  Callback extends (...args: any[]) => any,
->(): AsyncHook<Callback> {
-  const preGroup: Callback[] = [];
-  const postGroup: Callback[] = [];
-  const defaultGroup: Callback[] = [];
-
-  const tap = (cb: Callback | HookDescriptor<Callback>) => {
-    if (isFunction(cb)) {
-      defaultGroup.push(cb);
-    } else if (cb.order === 'pre') {
-      preGroup.push(cb.handler);
-    } else if (cb.order === 'post') {
-      postGroup.push(cb.handler);
-    } else {
-      defaultGroup.push(cb.handler);
-    }
-  };
-
-  const call = async (...args: Parameters<Callback>) => {
-    const params = args.slice(0) as Parameters<Callback>;
-    const callbacks = [...preGroup, ...defaultGroup, ...postGroup];
-
-    for (const callback of callbacks) {
-      const result = await callback(...params);
-
-      if (result !== undefined) {
-        params[0] = result;
-      }
-    }
-
-    return params;
-  };
-
-  return {
-    tap,
-    call,
-  };
-}
