+++ packages/shared/src/index.ts
@@ -1,11 +1,9 @@
 export * from './constants';
-export * from './createHook';
 export * from './utils';
 export * from './fs';
 export * from './getBrowserslist';
 export * from './logger';
 export * from './mergeChainedOptions';
-export * from './mergeRsbuildConfig';
 export * from './types';
 export * from './config';
 export * from './chain';
